conda_requirements.txt
rich
requests
python>=3.12
python-dotenv
jinja2
beautifulsoup4
